alphacore/backend/requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
Django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-filter==23.5
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0

# ── AWS DynamoDB ──────────────────────────────────────────────────────────────
boto3==1.34.34

# ── Auth (AlphaCore only) ─────────────────────────────────────────────────────
django-allauth==0.61.1
requests==2.31.0
